refactor to extract forms
